WiP service refactoring
