+++ src/config.ts
@@ -125,11 +125,13 @@ const defaultConfig: IoDevServerConfig = {
     },
     national: 5,
     local: 5,
-    includeSiciliaVola: true,
-    includeCgn: true,
-    includeCdc: true,
-    includePn: false,
-    includeFci: true,
+    specialServices: {
+      siciliaVola: true,
+      cgn: true,
+      cdc: true,
+      pn: false,
+      fci: true,
+    },
     // it has partially effect (pr welcome)
     allowRandomValues: true
   },
